refactor(e2e): convert CrossChainMessagingTest to a single-node harness
Move e2e_cross_chain_messaging/ into the new single-node/cross-chain/ folder
and reconcile CrossChainMessagingTest onto SingleNodeTestContext. The harness
now extends the base and builds its environment with its own setup() opts
(optional prover node, EpochTestSettler auto-proving otherwise), then calls
hydrateFromContext to reuse the base rollup / epoch-cache / chain-monitor /
node-tracking / teardown machinery, keeping the CrossChainTestHarness domain
object and the L1 inbox/outbox handles on the harness. The shared
cross_chain_test_harness.ts stays in src/shared/ (it is also used by
bench/client_flows and shared/uniswap_l1_l2); the moved files' import paths
are repointed accordingly. Adds the single-node/cross-chain/ CI glob to both
bootstrap arrays, repoints the per-test timeout case and the
.test_patterns.yml entries, and applies the .parallel suffix to the multi-it
files.
