+++ .test_patterns.yml
@@ -123,22 +123,22 @@ tests:
     error_regex: "✕ Repay"
     owners:
       - *lasse
-  - regex: "src/e2e_cross_chain_messaging/l1_to_l2"
+  - regex: "src/single-node/cross-chain/l1_to_l2"
     owners:
       - *palla
-  - regex: "src/e2e_cross_chain_messaging/token_bridge_private"
+  - regex: "src/single-node/cross-chain/token_bridge_private"
     error_regex: "✕ Claim secret is enough to consume the message"
     owners:
       - *lasse
-  - regex: "src/e2e_cross_chain_messaging/token_bridge_public"
+  - regex: "src/single-node/cross-chain/token_bridge_public"
     error_regex: "✕ Someone else can mint funds to me on my behalf"
     owners:
       - *lasse
-  - regex: "src/e2e_cross_chain_messaging/token_bridge_public"
+  - regex: "src/single-node/cross-chain/token_bridge_public"
     error_regex: "✕ Publicly deposit funds"
     owners:
       - *lasse
-  - regex: "src/e2e_cross_chain_messaging/token_bridge_failure_cases"
+  - regex: "src/single-node/cross-chain/token_bridge_failure_cases"
     error_regex: "✕ Can't claim funds"
     owners:
       - *lasse
